Added a draft normal FFT
Lots of changes involved.
~ changed to !
Asynchronous negative logic reset was added
The coefficient generation was pulled out of the FFT-Stage logic
The FFT-stages were collapsed into a single stage, parameterized by the main
  FFT module
The hard-multiply choice is now controlled by a parameter as well.
A couple more modules were added, to facilitate the single-clock logic.
"single_clock" is really a misnomer.  It doesn't so much mean "single clock" as
  it means one sample per clock.  This may yet need to be changed in the
  future for better readability.

The biggest problem so far: the test bench has yet to prove this new code does
(or does not) work.